+++ tests/test_arithmetics.rs
@@ -1,9 +1,8 @@
-#![feature(proc_macro, specialization)]
+#![feature(use_extern_macros, specialization)]
 
 extern crate pyo3;
 
 use pyo3::prelude::*;
-use pyo3::{pyclass, pyproto};
 
 #[macro_use]
 mod common;