+++ tests/test_class_basics.rs
@@ -1,10 +1,8 @@
-#![feature(proc_macro, specialization)]
+#![feature(use_extern_macros, specialization)]
 
 extern crate pyo3;
 
 use pyo3::prelude::*;
-
-use pyo3::pyclass;
 
 #[macro_use]
 mod common;