+++ tests/test_dunder.rs
@@ -1,14 +1,10 @@
-#![feature(proc_macro, specialization)]
+#![feature(use_extern_macros, specialization)]
 
 extern crate pyo3;
 
 use pyo3::ffi;
 use pyo3::prelude::*;
 use std::{isize, iter};
-
-use pyo3::pyclass;
-use pyo3::pymethods;
-use pyo3::pyproto;
 
 #[macro_use]
 mod common;