+++ tests/test_getter_setter.rs
@@ -1,12 +1,9 @@
-#![feature(proc_macro, specialization)]
+#![feature(use_extern_macros, specialization)]
 
 extern crate pyo3;
 
 use pyo3::prelude::*;
 use std::isize;
-
-use pyo3::pyclass;
-use pyo3::pymethods;
 
 #[macro_use]
 mod common;