+++ tests/test_inheritance.rs
@@ -1,11 +1,9 @@
-#![feature(proc_macro, specialization)]
+#![feature(use_extern_macros, specialization)]
 
 extern crate pyo3;
 
 use pyo3::prelude::*;
 use std::isize;
-
-use pyo3::{pyclass, pymethods};
 
 #[macro_use]
 mod common;